+++ crates/miden-objects/src/note/script.rs
@@ -4,10 +4,7 @@ use core::fmt::Display;
 use super::Felt;
 use crate::{
     NoteError, PrettyPrint, Word,
-    assembly::{
-        Assembler, Parse,
-        mast::{MastForest, MastNodeId},
-    },
+    assembly::mast::{MastForest, MastNodeId},
     utils::serde::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable},
     vm::Program,
 };
@@ -35,18 +32,6 @@ impl NoteScript {
             entrypoint: code.entrypoint(),
             mast: code.mast_forest().clone(),
         }
-    }
-
-    /// Returns a new [NoteScript] compiled from the provided source code using the specified
-    /// assembler.
-    ///
-    /// # Errors
-    /// Returns an error if the compilation of the provided source code fails.
-    pub fn compile(source_code: impl Parse, assembler: Assembler) -> Result<Self, NoteError> {
-        let program = assembler
-            .assemble_program(source_code)
-            .map_err(NoteError::NoteScriptAssemblyError)?;
-        Ok(Self::new(program))
     }
 
     /// Returns a new [NoteScript] deserialized from the provided bytes.
@@ -208,14 +193,15 @@ impl Display for NoteScript {
 
 #[cfg(test)]
 mod tests {
-    use super::{Assembler, Felt, NoteScript, Vec};
-    use crate::testing::note::DEFAULT_NOTE_CODE;
+    use super::{Felt, NoteScript, Vec};
+    use crate::{assembly::Assembler, testing::note::DEFAULT_NOTE_CODE};
 
     #[test]
     fn test_note_script_to_from_felt() {
         let assembler = Assembler::default();
         let tx_script_src = DEFAULT_NOTE_CODE;
-        let note_script = NoteScript::compile(tx_script_src, assembler).unwrap();
+        let program = assembler.assemble_program(tx_script_src).unwrap();
+        let note_script = NoteScript::new(program);
 
         let encoded: Vec<Felt> = (&note_script).into();
         let decoded: NoteScript = encoded.try_into().unwrap();