+++ crates/miden-objects/src/transaction/tx_args.rs
@@ -4,12 +4,11 @@ use alloc::{
     vec::Vec,
 };
 
-use assembly::{Assembler, Parse};
 use miden_crypto::merkle::InnerNodeInfo;
 
 use super::{AccountInputs, Felt, Word};
 use crate::{
-    EMPTY_WORD, MastForest, MastNodeId, TransactionScriptError,
+    EMPTY_WORD, MastForest, MastNodeId,
     note::{NoteId, NoteRecipient},
     utils::serde::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable},
     vm::{AdviceInputs, AdviceMap, Program},
@@ -272,21 +271,6 @@ impl TransactionScript {
     /// Returns a new [TransactionScript] instantiated with the provided code.
     pub fn new(code: Program) -> Self {
         Self::from_parts(code.mast_forest().clone(), code.entrypoint())
-    }
-
-    /// Returns a new [TransactionScript] compiled from the provided source code using the specified
-    /// assembler.
-    ///
-    /// # Errors
-    /// Returns an error if the compilation of the provided source code fails.
-    pub fn compile(
-        source_code: impl Parse,
-        assembler: Assembler,
-    ) -> Result<Self, TransactionScriptError> {
-        let program = assembler
-            .assemble_program(source_code)
-            .map_err(TransactionScriptError::AssemblyError)?;
-        Ok(Self::new(program))
     }
 
     /// Returns a new [TransactionScript] instantiated from the provided MAST forest and entrypoint.