+++ crates/miden-testing/tests/lib.rs
@@ -4,13 +4,13 @@ mod auth;
 mod scripts;
 mod wallet;
 
-use miden_lib::transaction::TransactionKernel;
+use miden_lib::utils::ScriptBuilder;
 use miden_objects::{
     Word, ZERO,
     account::AccountId,
     asset::FungibleAsset,
     crypto::utils::Serializable,
-    note::{Note, NoteAssets, NoteInputs, NoteMetadata, NoteRecipient, NoteScript, NoteType},
+    note::{Note, NoteAssets, NoteInputs, NoteMetadata, NoteRecipient, NoteType},
     testing::account_id::ACCOUNT_ID_SENDER,
     transaction::{ExecutedTransaction, ProvenTransaction},
 };
@@ -81,8 +81,7 @@ pub fn get_note_with_fungible_asset_and_script(
 ) -> Note {
     use miden_objects::note::NoteExecutionHint;
 
-    let assembler = TransactionKernel::assembler().with_debug_mode(true);
-    let note_script = NoteScript::compile(note_script, assembler).unwrap();
+    let note_script = ScriptBuilder::default().compile_note_script(note_script).unwrap();
     let serial_num = Word::from([1, 2, 3, 4u32]);
     let sender_id = AccountId::try_from(ACCOUNT_ID_SENDER).unwrap();